Move timer to main dashboard, replacing old timer, then style it alongside, other components on the site
