+++ frontend/src/App.jsx
@@ -7,7 +7,6 @@ import Header from './components/Header'
 import Login from './pages/Login'
 import Register from './pages/Register'
 import Profile from './pages/Profile'
-import TestTimer from './pages/TestTimer'
 
 function App() {
   return (
@@ -21,7 +20,6 @@ function App() {
           <Route path='/register' element={<Register />} />
           <Route path="/about" element={<About />} />
           <Route path="/profile" element={<Profile />} />
-          <Route path="/test-timer" element={<TestTimer />} />
         </Routes>
       </div>
     </Router>
